+++ libopenmpt/libopenmpt_cxx.cpp
@@ -312,6 +312,9 @@ std::int32_t module::get_current_speed() const {
 std::int32_t module::get_current_tempo() const {
 	return impl->get_current_tempo();
 }
+double module::get_current_tempo2() const {
+	return impl->get_current_tempo2();
+}
 std::int32_t module::get_current_order() const {
 	return impl->get_current_order();
 }